+++ src/tcp_server.cpp
@@ -29,7 +29,7 @@ server::server(scheduler& sched, const unsigned int tcp_port)
 #else
     m_ssl_context(0),
 #endif
-    m_endpoint(asio::ip::tcp::v4(), tcp_port), m_ssl_flag(false), m_is_listening(false)
+    m_endpoint(asio::ip::tcp::v4(), (unsigned short)tcp_port), m_ssl_flag(false), m_is_listening(false)
 {}
 
 server::server(scheduler& sched, const asio::ip::tcp::endpoint& endpoint)
@@ -53,7 +53,7 @@ server::server(const unsigned int tcp_port)
 #else
     m_ssl_context(0),
 #endif
-    m_endpoint(asio::ip::tcp::v4(), tcp_port), m_ssl_flag(false), m_is_listening(false)
+    m_endpoint(asio::ip::tcp::v4(), (unsigned short)tcp_port), m_ssl_flag(false), m_is_listening(false)
 {}
 
 server::server(const asio::ip::tcp::endpoint& endpoint)
